requirements.txt
amqp==5.0.5
asgiref==3.3.1
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.5
#django-celery-beat==2.2.0
django-environ==0.4.5
django-timezone-field==4.1.2
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
gunicorn==20.0.4
idna==2.10
itypes==1.2.0
Jinja2==2.11.3
kombu==5.0.2
MarkupSafe==1.1.1
prompt-toolkit==3.0.17
psycopg2-binary==2.8.5
PyJWT==2.0.1
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
uritemplate==3.0.1
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.1
django-cors-headers==3.7.0
redis==3.5.3
django-extensions==3.1.1
django-filter==2.4.0

#dnspython
django-celery-beat==2.2.0


